mouse controlled selection
This required a bit of a refactor because the month widget itself needed
to keep track of the click events so I had to make it properly implement
bubbletea.Model. Once that was arranged the zone library is very easy to
use and seems to have reasonable enough performance.

Currently, the selection is stored in the month widget itself, but this
isn't super ideal as I add more months. I think I'll need to make a sort
of "month controller" which can create and destroy the more simply month
widgets and keeps track of the selection state. Perhaps I could use
bubbletea.Cmd to pass the selection up to the controller.