display multiple months
This patch creates a new "calendar" element which is now used in the
main model instead of the "currentMonth" field. The calendar holds a
slice of month elements which is creates and deletes to match the
current window size.

For now, this means we can either display a single centered month or
three months stacked vertically, but in the future this pattern can be
expanded to show a full year grid or perhaps even more.