create config file
I've created a config file format (TOML based) which is used to
configure the colors in the UI. In the future I will also use it to
configure the path to your journal and override the default EDITOR.

I decided not to write man pages quite yet since the spec will still
likely change before the core features of the program are finished.
However, I did update the README to better represent the current state
of the program as it's now a concrete usable tool rather than an idea.