create preview element
I've created a preview element which is displayed when the window is
wider than 40 columns + 20 columns for the month elements. It also
stretches out to a max width of 80 characters, centering itself after
that. These values are of course configurable.

The preview widget itself shows a text file for the date using a
configurable path containing a date pattern. The default configured
NotePath is this:
$HOME/.local/share/calendar/2006-01-02.md

The path should describe where the note for the following date would be
stored:
January 2, 15:04:05, 2006, in time zone seven hours west of GMT

When the preview is open, you can press "tab" to focus is and then you
will be able to scroll long notes. Pressing tab again will return your
focus to the month widgets.