fix calendar centering when preview is hidden
Previously, the calendar was centered such that the bottom of the
calendar had the extra blank line when the window had an odd number of
rows. This resulted in it looking wonky since month month the last line
is blank already. Additionally, it was different from the behaviour of
the calendar with the preview window shown.