use NoteDir instead of NotePath
Credit to j3s.sh for writing this patch! Having the config for the date
layout use the same config option as the calendar directory meant that
you couldn't have any "unrelated" numbers in your path. So if your local
username was j3s it hilariously converted it to j6s and failed to open
your notes.

Perhaps in the future, we could add a separate config option back in for
changing the date format, but this works fine for now.